ConfirmHouseResize -> ConfirmHouseResizeGump and cleanup
